add/reserveページ，displayページ
